Disable I18n Alchemy for numeric fields (only use it for date/time/tod fields)
This is because I18n alchemy leads to problems when using the new HTML5 number fields.
These fields expect a non-localized value, and localize it for you. This change breaks
nice number formatting for older browsers, but that is something we can live with (and
could potentially be fixed using JavaScript).

This commit also cleans some of the I18n parsers and the way they are loaded.