Remove Reservation#slack_(before|after)_overlapping methods
(returning the reservation with slack before or after overlap is not necessary any more.
Or actually it was never necessary, since when we now there is slack overlap, we can
simply return the previous or next reservation as the reservation with slack overlap.
The slack_(before|after)_overlapping? methods are kept)