Bugfix in Reservation.by_date_domain
(we now consider the to time exclusive, just as with everything else in the application.
this fixes a bug in the reservations form where an entity was viewed as unavailable when
the reservation ended at the same time another reservation began. This is however not a
problem, since the ends_at time is (should be) viewed exclusive)