Add Reservations#full_instance_name method to replace #one_line_summary and local solutions
This method gives a standardised solution to display a Reservation as shortly (i.e. in one line).
It takes several options to control the final output. This method replaces the various other/local
implementation that all show a Reservation slightly different. With this method the formatting
of a Reservation will be consistent throughout the whole application