Add acts_as_list gem and refactor all existing sorting to it
Preparation for the tabs separation rewrite.

This commit adds the acts_as_list gem and renames the current #index and #weight
columns to #position, following the conventions of this gem. It also refactors
the way the position of stickies is set, using a more general approach which can
be used for entity type properties and reservation statuses as well later.